+++ CHANGELOG.md
@@ -311,6 +311,7 @@
 - Added a /bedrockparticle command to run Snowstorm-format particle effects.
 - Added data for Dipplin, Fezandipiti, Munkidori, Ogerpon, Okidogi, Poltchageist and Sinistcha.
 - Added additional nickname trigger "Grumm" for Inkay's evolution.
+- Added a animation and particles to Pokemon evolving while in the overworld.
 
 ### Changes
 - Using Potions, Status Heals, Ethers, and Antidotes will now return a glass bottle